Use MakeIdentityReturnMap across codebase; widen Hashset hash to size_t; revert CUDD cache ratio
- Replace 33 manual identity return map constructions with MakeIdentityReturnMap(k)
  across cflobdd_node.cpp, multiplication_crt.cpp, matrix1234_node.cpp,
  vector_node.cpp, cflobdd_top_node_int.cpp
- Move MakeIdentityReturnMap cache to function-local statics to prevent
  static initialization order issues
- Move setHashCheck() inside isCanonical guard in Canonicalize() to skip
  redundant hash computation on already-canonical maps
- Widen Hashset::GetHash return and Lookup/Insert hash params from
  unsigned int to size_t; remove premature % numBuckets from GetHash
  to decouple raw hash from bucket count (enables future table resizing)
- Revert DD_MAX_CACHE_TO_SLOTS_RATIO from 1024 back to standard value 4
  in cudd-3.0.0/cudd/cuddInt.h for fair CFLOBDD-vs-CUDD benchmarking
- Add multmod-all target and extend mult-tests in examples/Makefile

+++ CFLOBDD/cross_product.cpp
@@ -254,7 +254,7 @@ int PairProductMapHandle::Lookup(intpair& p)
 void PairProductMapHandle::Canonicalize()
 {
   PairProductMapBody *answerContents;
-  unsigned int hash = PairProductMapBody::canonicalPairProductMapBodySet->GetHash(mapContents);
+  size_t hash = PairProductMapBody::canonicalPairProductMapBodySet->GetHash(mapContents);
   answerContents = PairProductMapBody::canonicalPairProductMapBodySet->Lookup(mapContents, hash);
   if (answerContents == NULL) {
     PairProductMapBody::canonicalPairProductMapBodySet->Insert(mapContents, hash);
@@ -795,7 +795,7 @@ int TripleProductMapHandle::Lookup(inttriple t)
 void TripleProductMapHandle::Canonicalize()
 {
   TripleProductMapBody *answerContents;
-  unsigned int hash = TripleProductMapBody::canonicalTripleProductMapBodySet->GetHash(mapContents);
+  size_t hash = TripleProductMapBody::canonicalTripleProductMapBodySet->GetHash(mapContents);
   answerContents = TripleProductMapBody::canonicalTripleProductMapBodySet->Lookup(mapContents, hash);
   if (answerContents == NULL) {
     TripleProductMapBody::canonicalTripleProductMapBodySet->Insert(mapContents, hash);
